Change Value to be i16 instead of i32
It doesn't _need_ to be i32, and this way a TT entry takes up less space.

+++ goldfish/uci/src/types.rs
@@ -66,14 +66,14 @@ enum OptionType {
 
 #[derive(Debug)]
 pub enum InfoPart<'a> {
-    Cp(i32),
+    Cp(i16),
     CurrLine(&'a [ChessMove]),
     CurrMove(ChessMove),
     CurrMoveNumber(usize),
     Depth(usize),
     HashFull(usize),
     LowerBound(i32),
-    Mate(i32),
+    Mate(i16),
     MultiPv(usize),
     Nodes(usize),
     Nps(usize),
